+++ purchase_order_line_sequence/__init__.py
@@ -5,4 +5,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import models
-from .init_hooks import post_init_hook
+from .hooks import post_init_hook
